removed extra intersection correction

+++ srcs/utils/utils.c
@@ -6,7 +6,7 @@
 /*   By: hwong <hwong@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/18 19:12:31 by hwong             #+#    #+#             */
-/*   Updated: 2023/05/03 18:44:40 by hwong            ###   ########.fr       */
+/*   Updated: 2023/05/09 14:45:03 by hwong            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,8 +43,8 @@ int	is_in_wall( t_game *game, int x, int y )
 	int	map_x;
 	int	map_y;
 
-	map_x = (int)(x / CELL_SIZE);
-	map_y = (int)(y / CELL_SIZE);
+	map_x = (int)(x / (double)CELL_SIZE);
+	map_y = (int)(y / (double)CELL_SIZE);
 	if (map_x < 0 || map_y < 0 || map_x >= game->msize.x
 		|| map_y >= game->msize.y)
 		return (1);
